requirements.txt
email-validator
flask
flask-cors
flask-migrate
flask-sqlalchemy
flask-wtf
python-dotenv
werkzeug
wtforms